Added Documentation
Added a design goals doc to list what the goals are for the project as well as current things that are being worked on and an explanation as to why particular parts were chosen. Also added BOM to keep track of  what parts were used, where to source them, and how much they cost.