Pressure Readings and PID tuning
Can now read the pressure from the pressure sensor properlly. (in PSI) I believe the pressure sensor is als o properly calibrated, but the PID loop needs a lot of work. I'e realised that the PID target value will need to change depending on whether or not the valves are open. Seeing as blood pressure is measured by a relaxed and a tense heart rate (120/60) there will have to be a PID loop for each of these targets.