Added changing blood pressure and an extended running average
Blood pressure PID target now changes depending on the solenoid valve state. Also extended the running average to measure the pressure to 15 values for better accuracy. Retuned PID loop so it works MUCH better. The whole thing still needs a lot of work but the system might be ready for primetime.